[PROD-101] webfe: add gRPC-Web
This CL integrates webfe in our Bazel build system, and integrate
protobuf and gRPC-Web to the js webpack. Built asset is also bundled to
our aggregator golang binary so it's easy to deploy. A login RPC call is
added as an example. Use `-c dbg` for a debug build (unminified
javascript with sourcemap).

TODO: refactor all starlark occurences of config_setting debug_build to be
unified under the root BUILD file. This will be done in a subsequent CL
to avoid merge conflicts.

Reviewed-by: Chester Pang <i@bopa.ng>
Tested-by: Chester Pang <i@bopa.ng>
GitOrigin-RevId: 87eda0134afbec7b8040065903f2f0f4de7c83ec