aggregator: optimize algorithm and fix calculation bugs
It now calculates app-usage-intervals on the fly and store them to db.
An in-memory device-state is used to facilitate the device state
calculation (TODO: recover device-state upon server rebooting). An
event heap is added to support out-of-order events, with corresponding
unit tests.

For the querying RPC, we no longer loop through raw events - but our
pre-calcualted intervals instead. This also fixes several bugs in stats
calculation in the previous implementation.

Reviewed-by: Chester Pang <i@bopa.ng>
Tested-by: Chester Pang <i@bopa.ng>
GitOrigin-RevId: 95ab1857084bc5671ddbdc60a565076effa6940b