reporter: refactor is_tracking to core module
Plat code no longer needs to keep track of whether it's currently
tracking or not. A RWMutex is added so that if core module receives a
WindowSwitch event after tracking is stopped, that event will be
silently dropped.

Locks are still kept in plat code around starting and stopping tracking
(locks for windows were missing and are added), so we don't have race
condition for adding/removing OS-specific hooks if the user starts/stops
tracking too frequently.

Reviewed-by: Chester Pang <i@bopa.ng>
Tested-by: Adam Yi <i@adamyi.com>
GitOrigin-RevId: c8c03abf00f77284c3625eacd8e195b88326c814