webfe: display limited number of items in histogram
Allow user to configure the number items (stacked bars)
displayed in the histogram, by default it's 10.
In non-fullscreen mode, we display either 5 or the user's configured
number of items if it's smaller.
Items are ranked by their total time in the histogram and
any items other than the displayed ones will be put
in a separate category "Other"

Reviewed-by: Adam Yi <i@adamyi.com>
Tested-by: Chester Pang <i@bopa.ng>
GitOrigin-RevId: e6a97da2166ef5d5af38f8f5f80f3cc23d0e630d